refactore: move utils store package
