+++ cmd/cli/kubectl-kyverno/store/contextloader.go
